backend_python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
# psycopg2-binary==2.9.10  # For PostgreSQL
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.1
stripe==12.4.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
alembic==1.13.1
